// shared/schema.d.ts
export declare const adminStatsFilterSchema: z.ZodObject<{
    hours: z.ZodDefault<z.ZodOptional<z.ZodNumber>>;
    channel: z.ZodOptional<z.ZodEnum<["email", "sms"]>>;
    purpose: z.ZodOptional<z.ZodEnum<["signup", "login", "reset"]>>;
}, "strip", z.ZodTypeAny, {
    hours: number;
    channel?: "email" | "sms" | undefined;
    purpose?: "signup" | "login" | "reset" | undefined;
}, {
    channel?: "email" | "sms" | undefined;
    purpose?: "signup" | "login" | "reset" | undefined;
    hours?: number | undefined;
}>;